+++ hw07-ATM_department/src/main/java/ATMDepartment/Command/Command.java
@@ -2,7 +2,7 @@
  * Created by Ilya Rogatkin, May 2020
  */
 
-package ATM_department.Command;
+package ATMDepartment.Command;
 
 public interface Command {
     void execute();
