+++ hw07-ATM_department/src/main/java/ATMDepartment/chain/ATMChecker.java
@@ -2,24 +2,20 @@
  * Created by Ilya Rogatkin, May 2020
  */
 
-package ATM_department.chain;
+package ATMDepartment.chain;
 
-import ATM.ATMCityBank;
-
-public class ATMInitializer extends ATMProcessor {
+public class ATMChecker extends ATMProcessor {
     @Override
     protected void processInternal(ATMDepartment department) {
         try {
-            final int NUMBER_OF_ATMS = 5;
-            department.initializeATMs(NUMBER_OF_ATMS);
+            department.checkSumInATM();
         } catch (Exception err) {
             System.out.println(err.getMessage());
         }
-
     }
 
     @Override
     public String getProcessorName() {
-        return "Инициализация банкоматов";
+        return "Проверка данных и загруженных купюр";
     }
 }
