+++ hw07-ATM_department/src/main/java/ATMDepartment/chain/ATMInitializer.java
@@ -2,20 +2,22 @@
  * Created by Ilya Rogatkin, May 2020
  */
 
-package ATM_department.chain;
+package ATMDepartment.chain;
 
-public class ATMChecker extends ATMProcessor {
+public class ATMInitializer extends ATMProcessor {
     @Override
     protected void processInternal(ATMDepartment department) {
         try {
-            department.checkSumInATM();
+            final int NUMBER_OF_ATMS = 5;
+            department.initializeATMs(NUMBER_OF_ATMS);
         } catch (Exception err) {
             System.out.println(err.getMessage());
         }
+
     }
 
     @Override
     public String getProcessorName() {
-        return "Проверка данных и загруженных купюр";
+        return "Инициализация банкоматов";
     }
 }
