+++ hw07-ATM_department/src/main/java/ATMDepartment/chain/ATMProcessor.java
@@ -2,7 +2,7 @@
  * Created by Ilya Rogatkin, May 2020
  */
 
-package ATM_department.chain;
+package ATMDepartment.chain;
 
 abstract class ATMProcessor implements Handler{
     private Handler next;
