+++ hw07-ATM_department/src/main/java/ATMDepartment/memento/Memento.java
@@ -2,7 +2,7 @@
  * Created by Ilya Rogatkin, May 2020
  */
 
-package ATM_department.memento;
+package ATMDepartment.memento;
 
 class Memento {
     private final State state;
