+++ hw07-ATM_department/src/main/java/ATMDepartment/observer/Listener.java
@@ -2,7 +2,7 @@
  * Created by Ilya Rogatkin, May 2020
  */
 
-package ATM_department.observer;
+package ATMDepartment.observer;
 
 public interface Listener {
     void onUpdate();
