+++ hw07-ATM_department/src/main/java/ATMDepartment/observer/Producer.java
@@ -2,9 +2,7 @@
  * Created by Ilya Rogatkin, May 2020
  */
 
-package ATM_department.observer;
-
-import ATM_department.Command.Command;
+package ATMDepartment.observer;
 
 import java.util.ArrayList;
 import java.util.List;
